pyproject.toml
[project]
name = "base-image-plus"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=11.0.0",
    "toml>=0.10.2",
]
